fixes151b: int32_t storage for grid-col-span (no struct padding)
A uint8_t between two int32_t fields in css_computed_style_i creates 3
bytes of compiler padding. While the field itself is byte-deterministic
when written by my cascade, the padding bytes are written by NOTHING --
they retain whatever value calloc set (usually 0) until a cascade
compose path constructs a new style some other way.

Same failure mode as fixes117's inline track array: arena interning's
memcmp(&a->i, &b->i, sizeof(_i)) compares all bytes including padding;
two logically-equal styles with different padding patterns hash into
separate intern buckets; the duplicate styles get queued for destroy;
css_computed_style_destroy walks the duplicate, frees outer-struct
pointers, then a second walk frees them again -> use-after-free.

Fix: change the field to int32_t so it is self-aligning. No padding
bytes. memcmp is bullet-proof. Public accessor still returns uint8_t
for layout consumer clarity, with defensive clamp.

+++ browser/libcss/src/select/properties/s_macsurf_grid_col_span.c
@@ -40,7 +40,7 @@ css_error css__cascade_macsurf_grid_col_span(uint32_t opv,
 	if (css__outranks_existing(getOpcode(opv), isImportant(opv), state,
 			getFlagValue(opv))) {
 		state->computed->i.macsurf_grid_col_span =
-				is_set ? (uint8_t)span : 0;
+				is_set ? span : 0;
 	}
 
 	return CSS_OK;
@@ -77,7 +77,7 @@ css_error css__compose_macsurf_grid_col_span(
 	/* col-span is not inherited; child wins unless child is unset
 	 * (0) and parent is set, in which case fall back to parent.
 	 * Matches the pattern used by macsurf_grid_rows. */
-	uint8_t v = child->i.macsurf_grid_col_span;
+	int32_t v = child->i.macsurf_grid_col_span;
 	if (v == 0) v = parent->i.macsurf_grid_col_span;
 	result->i.macsurf_grid_col_span = v;
 	return CSS_OK;
